requirements.txt
langchain
python-dotenv
langchain-community
ipykernel
PyPDF2
tiktoken
huggingface_hub
pinecone-client
pinecone
google-generativeai
ctransformers
langchain-huggingface
sentance-transformer

pyreadline3
tensorflow==2.11.0
pdfplumber
fastapi
uvicorn
BeautifulSoup4